+++ lib/index.tsx
@@ -1,7 +1,6 @@
 import * as isEqual from 'lodash.isequal';
 import * as qrGenerator from 'qrcode-generator';
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
 
 type EyeColor = string | InnerOuterEyeColor;
 type InnerOuterEyeColor = {
@@ -27,14 +26,14 @@ export interface IProps {
     logoWidth?: number;
     logoHeight?: number;
     logoOpacity?: number;
-    logoOnLoad?: () => void;
+    logoOnLoad?: (e: Event) => void;
     removeQrCodeBehindLogo?: boolean;
     logoPadding?: number;
     logoPaddingStyle?: 'square' | 'circle';
     eyeRadius?: CornerRadii | [CornerRadii, CornerRadii, CornerRadii];
     eyeColor?: EyeColor | [EyeColor, EyeColor, EyeColor];
     qrStyle?: 'squares' | 'dots' | 'fluid';
-    style?: object;
+    style?: React.CSSProperties;
     id?: string;
 }
 
@@ -45,7 +44,7 @@ interface ICoordinates {
 
 export class QRCode extends React.Component<IProps, {}> {
 
-    private canvas: React.RefObject<HTMLCanvasElement>;
+    private canvasRef = React.createRef<HTMLCanvasElement>();
 
     public static defaultProps: IProps = {
         value: 'https://reactjs.org/',
@@ -61,7 +60,29 @@ export class QRCode extends React.Component<IProps, {}> {
         logoPaddingStyle: 'square'
     };
 
-    private static utf16to8(str: string): string {
+    public download(fileType?: 'png' | 'jpg' | 'webp', fileName?: string) {
+        if (this.canvasRef.current) {
+            let mimeType;
+
+            switch (fileType) {
+                case 'jpg':
+                    mimeType = 'image/jpeg'; break;
+                case 'webp':
+                    mimeType = 'image/webp'; break;
+                case 'png':
+                default:
+                    mimeType = 'image/png'; break;
+            }
+
+            const url = this.canvasRef.current.toDataURL(mimeType, 1.0);
+            const link = document.createElement('a');
+            link.download = fileName ?? 'react-qrcode-logo';
+            link.href = url;
+            link.click();
+        }
+    }
+
+    private utf16to8(str: string): string {
         let out: string = '', i: number, c: number;
         const len: number = str.length;
         for (i = 0; i < len; i++) {
@@ -220,7 +241,6 @@ export class QRCode extends React.Component<IProps, {}> {
 
     constructor(props: IProps) {
         super(props);
-        this.canvas = React.createRef();
     }
 
     shouldComponentUpdate(nextProps: IProps) {
@@ -260,10 +280,10 @@ export class QRCode extends React.Component<IProps, {}> {
         const logoPadding = this.props.logoPadding ? +this.props.logoPadding : 0;
 
         const qrCode = qrGenerator(0, ecLevel);
-        qrCode.addData(QRCode.utf16to8(value));
+        qrCode.addData(this.utf16to8(value));
         qrCode.make();
 
-        const canvas: HTMLCanvasElement = ReactDOM.findDOMNode(this.canvas.current) as HTMLCanvasElement;
+        const canvas: HTMLCanvasElement = this.canvasRef?.current;
         const ctx: CanvasRenderingContext2D = canvas.getContext('2d');
 
         const canvasSize = size + (2 * quietZone);
@@ -312,8 +332,8 @@ export class QRCode extends React.Component<IProps, {}> {
                         let roundedCorners = [false, false, false, false]; // top-left, top-right, bottom-right, bottom-left
                         if ((row > 0 && !qrCode.isDark(row - 1, col)) && (col > 0 && !qrCode.isDark(row, col - 1))) roundedCorners[0] = true;
                         if ((row > 0 && !qrCode.isDark(row - 1, col)) && (col < length - 1 && !qrCode.isDark(row, col + 1))) roundedCorners[1] = true;
-                        if ((row < length -1 && !qrCode.isDark(row + 1, col)) && (col < length - 1 && !qrCode.isDark(row, col + 1))) roundedCorners[2] = true;
-                        if ((row < length -1 && !qrCode.isDark(row + 1, col)) && (col > 0 && !qrCode.isDark(row, col - 1))) roundedCorners[3] = true;
+                        if ((row < length - 1 && !qrCode.isDark(row + 1, col)) && (col < length - 1 && !qrCode.isDark(row, col + 1))) roundedCorners[2] = true;
+                        if ((row < length - 1 && !qrCode.isDark(row + 1, col)) && (col > 0 && !qrCode.isDark(row, col - 1))) roundedCorners[3] = true;
                         const w = (Math.ceil((col + 1) * cellSize) - Math.floor(col * cellSize));
                         const h = (Math.ceil((row + 1) * cellSize) - Math.floor(row * cellSize));
                         ctx.fillStyle = fgColor;
@@ -326,7 +346,7 @@ export class QRCode extends React.Component<IProps, {}> {
                             2 * Math.PI,
                             false);
                         ctx.closePath();
-                        ctx.fill(); 
+                        ctx.fill();
                         if (!roundedCorners[0]) ctx.fillRect(Math.round(col * cellSize) + offset, Math.round(row * cellSize) + offset, w / 2, h / 2)
                         if (!roundedCorners[1]) ctx.fillRect(Math.round(col * cellSize) + offset + Math.floor(w / 2), Math.round(row * cellSize) + offset, w / 2, h / 2)
                         if (!roundedCorners[2]) ctx.fillRect(Math.round(col * cellSize) + offset + Math.floor(w / 2), Math.round(row * cellSize) + offset + Math.floor(h / 2), w / 2, h / 2)
@@ -379,7 +399,7 @@ export class QRCode extends React.Component<IProps, {}> {
             if (enableCORS) {
                 image.crossOrigin = 'Anonymous';
             }
-            image.onload = () => {
+            image.onload = (e: Event) => {
                 ctx.save();
 
                 const dWidthLogo = logoWidth || size * 0.2;
@@ -413,7 +433,7 @@ export class QRCode extends React.Component<IProps, {}> {
                 ctx.drawImage(image, dxLogo + offset, dyLogo + offset, dWidthLogo, dHeightLogo);
                 ctx.restore();
                 if (logoOnLoad) {
-                    logoOnLoad();
+                    logoOnLoad(e);
                 }
             };
             image.src = logoImage;
@@ -422,12 +442,13 @@ export class QRCode extends React.Component<IProps, {}> {
 
     render() {
         const qrSize = +this.props.size + (2 * +this.props.quietZone);
-        return React.createElement('canvas', {
-            id: this.props.id ?? 'react-qrcode-logo',
-            height: qrSize,
-            width: qrSize,
-            style: { height: qrSize + 'px', width: qrSize + 'px' , ...this.props.style },
-            ref: this.canvas
-        });
+
+        return <canvas
+            id={this.props.id ?? 'react-qrcode-logo'}
+            height={qrSize}
+            width={qrSize}
+            style={{ height: qrSize + 'px', width: qrSize + 'px', ...this.props.style }}
+            ref={this.canvasRef}
+        />;
     }
 }